Move general code to specification jar, see #10154
