+++ components/specification/src/ome/specification/XMLMockObjects.java
@@ -2,7 +2,7 @@
  * unit.XMLMockFactory
  *
  *------------------------------------------------------------------------------
- *  Copyright (C) 2006-2010 University of Dundee. All rights reserved.
+ *  Copyright (C) 2006-2013 University of Dundee. All rights reserved.
  *
  *
  * 	This program is free software; you can redistribute it and/or modify
@@ -20,7 +20,7 @@
  *
  *------------------------------------------------------------------------------
  */
-package spec;
+package ome.specification;
 
 //Java imports
 import java.awt.Color;
