+++ components/specification/test/spec/schema/SchemaCurrent_TO_2003_FC_Test.java
@@ -25,10 +25,10 @@ import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
 //Application-internal dependencies
-import spec.AbstractTest;
-import spec.OmeValidator;
-import spec.XMLMockObjects;
-import spec.XMLWriter;
+import ome.specification.AbstractTest;
+import ome.specification.OmeValidator;
+import ome.specification.XMLMockObjects;
+import ome.specification.XMLWriter;
 
 import org.testng.Assert;
